refactor: Update helmrelease.yaml for pastebin app
This commit refactors the helmrelease.yaml file for the pastebin app. It updates the name from "microbin" to "pastebin" and adds an interval of 30 minutes. The chart version is also updated to 3.4.0. Additionally, it includes changes to the install and upgrade sections, adding remediation options and cleanup on fail. The values section is modified to use the "pastebin" controller and secret.

Update ks.yaml for pastebin app

This commit updates the ks.yaml file for the pastebin app. It adds a dependency on "longhorn" and removes the commented-out dependencies on "cluster-apps-external-secrets-stores". The path is set to "./kubernetes/apps/security/pastebin/app" and pruning is enabled.