✨ feat(minecraft): Add WorldEdit, LuckPerms and Distant Horizons
WorldEdit is a Minecraft map editor... that runs in-game! With
selections, schematics, copy and paste, brushes, and scripting.

LuckPerms is a permission plugin so that you can actually allow me to
use WorldEdit.

Distant Horizons can be used entirely client side, but the server side
component additionally allows the server to generate LODs, which is
cool.

Distant Horizons requires Fabric, so I've switched the server from Paper
to Fabric. Hopefully that won't break anything.

Also, I've added myself to ops 😉